+++ src/num/mem.h
@@ -4,7 +4,7 @@ extern void memcache_init(void);
 extern void memcache_destroy(void);
 extern void memcache_off(void);
 
-extern void* mem_device_malloc(long size, void* (*device_alloc)(size_t));
+extern void* mem_device_malloc(size_t size, void* (*device_alloc)(size_t));
 extern void mem_device_free(void* ptr, void (*device_free)(const void* x));
 extern void memcache_clear(void (*device_free)(const void* x));
